+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Support for the new encoding options in LibHeif versions 1.9.2, 1.10 and 1.11.
+
 ## v2.0.2 - 2020-12-29
 
 ### Fixed